Use useRef instead of useState and setState to store TinyCME content
There was a loop where the state is put into TinyCME's initialValue, and
when we update the content, the state is updated, updating initialValue,
triggering TinyCME to re-render. This resets the cursor to the
beginning, making it unusable.

This initial way broke TinyMCE as it was updated recently.

So instead of setState onEditorChange, we store the editor's content
onto the ref via useRef instead.

Change all instances where this editor is used to use ref instead of
state.