A loading state in SubscribeToProduct and a higher level loading state
When a button is click, we change that button's wording to say loading
cart/portal. The higher level loading state still disables the other
subscribe/update subscription buttons.

The useEffect in SubscribeToProduct basically resets loading if the
higher level loading is back to null. Basically if there's an error, we
want to re-enable all buttons.