Extend Exercise component to also handle AWL tests
Generally still using "exercise" as variables, but really it's test.

New route for editing a test - using "exerciseId" as the test ID

The Exercise component takes groupId as a param. Using this as a proxy
for whether it's an exercise or test, ie if groupId is null, it's a
test.

Update use of exercise or test depending on groupId.

Update redux and firestore data grabbing based on groupId.