Add instructions section to subscription page
Instruction section replaces the how it works section and move how it
works section below.

Also detect if there's "instructions" in the URL as a hash, so taht we
can emphasise the instructions section. Thinking of having a link to
this page from the home page with the instructions to be emphasised.

Drafted some instructions based on suggestions from Gerry.