Use React's context to handle auth and its consumption
Hopefully simplifying and optimising the loading and consumption of auth
data!

Hoisting the loading of this auth stuff into the Main component,
and the subsequent children can consume it using React's useContext
where necessary.

Should help reduce the boilerplate code to get this auth data...

Should reduce the number of loading too hopefully..