Create context for user info and current subscription
Can load in user info with the UID from the auth state! Yay! Loading
this in and then children can consume via the context.

Also loading in current subscription. Using this as the source of truth
for what user has subscribed.. Can remove the main field from the user
document, and remove the Cloud Function code that sets this on
subscription!

With these two pieces of info, simplified the useSubscription hook. No
need to account for whether the subscription info has been loaded or not
since it's being taken care of during the initial load.