Implement insitutional subscription
Show info about the tiered pricing, and update our cart and portal hook
to handle quantity. We're lucky here that the Stripe stuff accepts
quantity and will handle the price calculation correctly for us!

Also implemented access code. New access code collection with each doc
being just the UID of the user that purchased this multiple
subscription. The access code doc will have some normalised info we can
use to check for its validity and whether more users can be subscribed
to it.

Updated access checking for groups and exercises - also checks if the
associated access code is valid.

Updated our subscription Cloud Function to handle creating and updating
the new access code doc when a institutional subscription is purchased.
Also removed the old way of setting subscriptions on the User doc - now
checking for the latest subscription instead of checking for a value in
the doc.