Handle case where user doc is not present
Because we've configured Stripe to not create a customer on sign up, a
user document will not be present until user clicks one of the submit
buttons. So need to update code to account for user doc not being
present.

Main of changes lie in adding access code. No need for removal from
existing access code check if no user doc. And create a new user
document when providing an access code!