Update Route component prop and getting of params from URL
Pass element instead of component (Component in new router library).
There may be an optimisation for using the element prop..

Since we're instantiating components here, may be hard to provide the
required props. Most of the time, it's just passing through the match
object. Router provides a useParams and we can grab the required params
from it.

Params destructured may be empty, so need to default to empty string if
empty where it's getting consumed.