Update hook form usage
Destructure errors from formState

Register needs destructuring, and don't need to be passed as a ref prop
to the input element. The name attribute gets inserted as part of the
register destructuring, so no need for an explicit name attribute on the
input.

Rendering errors from the formState needs an explciit toString call